updated examples and json builder

+++ cleanup.js
@@ -3,6 +3,8 @@ var rmdir = require('rmdir');
 
 var rootSrc = 'examples';
 
+console.log('running example cleanup');
+
 fs.readdirSync(rootSrc).forEach(function(file) {
     if(isDirectory(rootSrc + '/' + file)) {
         secondLayer(rootSrc + '/' + file);
@@ -17,7 +19,9 @@ function secondLayer(path) {
             var lightningJS = path + '/' + file + '/static/js/lightning.min.js';
             rmdir(nodeModulesPath, function (err, dirs, files) {});
             rmdir(publicPath, function (err, dirs, files) {});
-            fs.unlinkSync(lightningJS);
+            if (fs.existsSync(lightningJS)) {
+                fs.unlinkSync(lightningJS);
+            }
         }
     });
 }
